Move scraper files to gridcoin directory
