+++ src/gridcoin/scraper/http.h
