+++ src/gridcoin/scraper/scraper.h
@@ -25,7 +25,7 @@
 #include "rpcprotocol.h"
 
 // See fwd.h for certain forward declarations that need to be included in other areas.
-#include "fwd.h"
+#include "gridcoin/scraper/fwd.h"
 #include "gridcoin/superblock.h"
 
 /*********************
