+++ src/gridcoin/scraper/scraper_net.h
