--- src/scraper/.gitignore
@@ -1,7 +1,0 @@
-userpass.dat
-scraper.time
-scraper.log
-GridcoinScraper2.pro*
-main.cpp
-test
-*.gz
